requirements.txt
streamlit
openai
pymupdf
python-dotenv
apify-client



-e .